Update Final Project Code
While I was doing this, I found that there is no way (or to my knowledge) how to make a buffer out of a density. So I think we should head in a direction where we make a buffer with a police station and have the user look at what kinds of crimes are happening in the buffer feature. We can keep the crime count for the funsies though. LMK if you see this. Making progress here:)